backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
google-genai
python-dotenv
numpy
